Move creating group from Utils
