+++ app/src/main/java/com/rebalance/utils/LogicUtils.kt
@@ -3,15 +3,6 @@ package com.rebalance.utils
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.util.Base64
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.snapshots.SnapshotStateMap
-import androidx.compose.ui.text.input.TextFieldValue
-import com.rebalance.PreferencesData
-import com.rebalance.backend.entities.ApplicationUser
-import com.rebalance.backend.entities.Expense
-import com.rebalance.backend.entities.ExpenseGroup
-import com.rebalance.backend.service.BackendService
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.FileOutputStream
@@ -20,23 +11,6 @@ import java.time.format.DateTimeFormatter
 
 fun getToday(): String {
     return LocalDate.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
-}
-
-
-
-fun createGroup(
-    groupCurrency: TextFieldValue,
-    groupName: TextFieldValue,
-    context: Context,
-    preferences: PreferencesData
-): ExpenseGroup? {
-    if (groupCurrency.text.length != 3 || groupName.text.isBlank()) {
-        alertUser("Fill in all fields!", context)
-        return null
-    }
-    val group = BackendService(preferences).createGroup(groupCurrency.text, groupName.text)
-    alertUser("Group was created!", context)
-    return group
 }
 
 fun compressImage(originalImage: Bitmap?, context: Context): Bitmap? {
